chore: CE-1120 Add Josh and Suha to users (plus bonus Cypress tweaks) (#712)
